pass hex files for test in environment, symlink test from project
